arch: decouple context module from infra (#178)
Split context.rs into domain and app layers:
- domain/context.rs: pure types (NodeKind, Node, MainBranch, ContextConfig,
  MaterializedMessage), pure functions (should_compact, default_main_path)
- app/context.rs: I/O operations (load, save, materialize) + re-exports

domain/context.rs has zero infra/app imports — only serde and std.

Closes #178
